move functions into proper folder
