+++ components/inputs/Select.tsx
@@ -3,7 +3,7 @@ import Router, { withRouter, SingletonRouter } from 'next/router';
 import { useQuery } from 'react-apollo';
 import styled from 'styled-components';
 import { DocumentNode } from 'graphql';
-import { urlBuilder } from '../../utils/url';
+import { urlBuilder } from '../../utils/functions/url';
 
 const StyledSelect = styled.select`
   font-family: var(--font-main);
